+++ JokEngine/InputBox.cpp
@@ -10,11 +10,11 @@ void InputBox::ReceiveInput(unsigned int input)
 {
 	if (isReadingInput)
 	{
-		if (maxCharacters <= text.size() && isTruncating)
+		if ((size_t)maxCharacters <= text.size() && isTruncating)
 		{
 			text.erase(0,1);
 		}
-		if (maxCharacters > text.size())
+		if ((size_t)maxCharacters > text.size())
 		{
 			wchar_t c = static_cast<wchar_t>(input);
 			text += toupper(input);
